+++ src/main/java/redstonedubstep/mods/vanishmod/misc/SoundSuppressionHelper.java
@@ -7,8 +7,10 @@ import java.util.Objects;
 import org.apache.commons.lang3.tuple.Pair;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.network.protocol.Packet;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.GameType;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.BlockHitResult;
@@ -16,10 +18,12 @@ import net.minecraft.world.phys.Vec3;
 import net.minecraft.world.phys.shapes.BooleanOp;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
+import redstonedubstep.mods.vanishmod.VanishConfig;
 import redstonedubstep.mods.vanishmod.VanishUtil;
 
 public class SoundSuppressionHelper {
 	private static final Map<ServerPlayer, Pair<BlockPos, Entity>> vanishedPlayersAndHitResults = new HashMap<>();
+	private static Pair<Packet<?>, Player> packetOrigin = null;
 
 	public static void updateVanishedPlayerMap(ServerPlayer player, boolean vanished) {
 		if (vanished)
@@ -43,6 +47,41 @@ public class SoundSuppressionHelper {
 		if (VanishUtil.isVanished(player))
 			vanishedPlayersAndHitResults.put(player, null);
 	}
+
+	public static void putSoundPacket(Packet<?> packet, Player player) {
+		packetOrigin = Pair.of(packet, player);
+	}
+
+	public static Player getPlayerForPacket(Packet<?> packet) {
+		return packetOrigin != null && packetOrigin.getLeft().equals(packet) ? packetOrigin.getRight() : null;
+	}
+
+	public static boolean shouldSuppressSoundEventFor(Player player, Level level, double x, double y, double z, Player forPlayer) {
+		return shouldSuppressSoundEventFor(player, level, new Vec3(x, y, z), forPlayer);
+	}
+
+	//Returns true if a vanished player directly produced the sound, or if it is determined that a vanished player was indirectly causing a sound, and that it thus should not be broadcast
+	public static boolean shouldSuppressSoundEventFor(Player player, Level level, Vec3 soundOrigin, Player forPlayer) {
+		if (VanishUtil.isVanished(player, forPlayer))
+			return true;
+
+		if (!VanishConfig.CONFIG.indirectSoundSuppression.get() || VanishUtil.canSeeVanishedPlayers(forPlayer))
+			return false;
+
+		return SoundSuppressionHelper.areVanishedPlayersAt(level, soundOrigin) || SoundSuppressionHelper.vanishedPlayerVehicleAt(level, soundOrigin) || SoundSuppressionHelper.vanishedPlayersInteractWith(level, new BlockPos(soundOrigin));
+	}
+
+	//Returns true if a vanished player directly produced the sound, or if it is determined that a vanished player was indirectly causing a sound, and that it thus should not be broadcast
+	public static boolean shouldSuppressSoundEventFor(Player player, Level level, Entity soundOrigin, Player forPlayer) {
+		if (VanishUtil.isVanished(player, forPlayer))
+			return true;
+
+		if (!VanishConfig.CONFIG.indirectSoundSuppression.get() || VanishUtil.canSeeVanishedPlayers(forPlayer))
+			return false;
+
+		return SoundSuppressionHelper.areVanishedPlayersAt(level, soundOrigin.position()) || SoundSuppressionHelper.isVanishedPlayerVehicle(soundOrigin) || SoundSuppressionHelper.vanishedPlayersInteractWith(level, soundOrigin);
+	}
+
 
 	public static boolean areVanishedPlayersAt(Level level, Vec3 pos) {
 		VoxelShape shape = Shapes.block().move(pos.x, pos.y, pos.z);