+++ src/main/java/redstonedubstep/mods/vanishmod/mixin/MixinServerGamePacketListenerImpl.java
@@ -1,5 +1,7 @@
 package redstonedubstep.mods.vanishmod.mixin;
 
+import java.util.List;
+
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -7,15 +9,20 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import com.mojang.authlib.GameProfile;
-
 import net.minecraft.network.protocol.Packet;
+import net.minecraft.network.protocol.game.ClientboundLevelEventPacket;
 import net.minecraft.network.protocol.game.ClientboundPlayerInfoPacket;
 import net.minecraft.network.protocol.game.ClientboundPlayerInfoPacket.Action;
+import net.minecraft.network.protocol.game.ClientboundPlayerInfoPacket.PlayerUpdate;
+import net.minecraft.network.protocol.game.ClientboundSoundEntityPacket;
+import net.minecraft.network.protocol.game.ClientboundSoundPacket;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.network.ServerGamePacketListenerImpl;
+import net.minecraft.world.phys.Vec3;
+import redstonedubstep.mods.vanishmod.VanishConfig;
 import redstonedubstep.mods.vanishmod.VanishUtil;
+import redstonedubstep.mods.vanishmod.misc.SoundSuppressionHelper;
 
 @Mixin(ServerGamePacketListenerImpl.class)
 public class MixinServerGamePacketListenerImpl {
@@ -24,19 +31,26 @@ public class MixinServerGamePacketListenerImpl {
 	@Shadow @Final
 	private MinecraftServer server;
 
-	//Prevent tab list information packet from being sent to other players if the player in the packet is vanished for the receiver.
-	//We don't filter latency packets, because they would break this method as there are multiple players in the "entries" argument of a latency update packet
-	//We also don't filter player removal packets, because this mod uses them to remove vanished players after their status has changed to be vanished,
+	//Filter any packets that we wish to not send to certain clients, mainly consisting of player info and sound packets.
+	//We don't filter player info removal packets, because this mod uses them to remove players after their status has changed to be vanished,
 	//and it can be done safely because not suppressing these packets does not break this mod (in: a player removal packet sent too much wouldn't break this mod as much as a player addition packet)
 	@Inject(method = "send(Lnet/minecraft/network/protocol/Packet;)V", at = @At("HEAD"), cancellable = true)
 	private void onSendPacket(Packet<?> packet, CallbackInfo callbackInfo) {
-		if (packet instanceof ClientboundPlayerInfoPacket infoPacket && infoPacket.getAction() != Action.UPDATE_LATENCY && infoPacket.getAction() != Action.REMOVE_PLAYER) {
-			GameProfile sentPlayer = !infoPacket.getEntries().isEmpty() ? infoPacket.getEntries().get(0).getProfile() : null;
+		if (packet instanceof ClientboundPlayerInfoPacket infoPacket && infoPacket.getAction() != Action.REMOVE_PLAYER) {
+			List<PlayerUpdate> vanishedPacketEntries = infoPacket.getEntries().stream().filter(p -> VanishUtil.isVanished(server.getPlayerList().getPlayer(p.getProfile().getId()), player)).toList();
 
-			if (sentPlayer != null) {
-				if (VanishUtil.isVanished(server.getPlayerList().getPlayer(sentPlayer.getId()), player))
-					callbackInfo.cancel();
-			}
+			if (vanishedPacketEntries.equals(infoPacket.getEntries()))
+				callbackInfo.cancel();
+			else if (!vanishedPacketEntries.isEmpty())
+				infoPacket.getEntries().removeAll(vanishedPacketEntries);
+		}
+		else if (VanishConfig.CONFIG.hidePlayersFromWorld.get()) {
+			if (packet instanceof ClientboundSoundPacket soundPacket && SoundSuppressionHelper.shouldSuppressSoundEventFor(SoundSuppressionHelper.getPlayerForPacket(soundPacket), player.level, soundPacket.getX(), soundPacket.getY(), soundPacket.getZ(), player))
+				callbackInfo.cancel();
+			else if (packet instanceof ClientboundSoundEntityPacket soundPacket && SoundSuppressionHelper.shouldSuppressSoundEventFor(SoundSuppressionHelper.getPlayerForPacket(soundPacket), player.level, player.level.getEntity(soundPacket.getId()), player))
+				callbackInfo.cancel();
+			else if (packet instanceof ClientboundLevelEventPacket soundPacket && SoundSuppressionHelper.shouldSuppressSoundEventFor(SoundSuppressionHelper.getPlayerForPacket(soundPacket), player.level, Vec3.atCenterOf(soundPacket.getPos()), player))
+				callbackInfo.cancel();
 		}
 	}
 }